+++ core/apis.go
@@ -107,6 +107,14 @@ func (s *administrationImpl) AdmUpdateOrganization(ID string, name string, reque
 	return s.app.admUpdateOrganization(ID, name, requestType, requiresOwnLogin, loginTypes, organizationDomains)
 }
 
+func (s *administrationImpl) AdmGetOrganizations() ([]model.Organization, error) {
+	return s.app.admGetOrganizations()
+}
+
+func (s *administrationImpl) AdmGetOrganization(ID string) (*model.Organization, error) {
+	return s.app.admGetOrganization(ID)
+}
+
 ///
 
 //encryptionImpl