+++ core/app_administration.go
@@ -208,6 +208,24 @@ func (app *application) admCreateOrganization(name string, requestType string, r
 	return organization, nil
 }
 
+func (app *application) admGetOrganization(ID string) (*model.Organization, error) {
+	organization, err := app.storage.GetOrganization(ID)
+	if err != nil {
+		return nil, log.WrapActionError(log.ActionGet, model.TypeOrganization, nil, err)
+	}
+
+	return organization, nil
+}
+
+func (app *application) admGetOrganizations() ([]model.Organization, error) {
+	getOrganization, err := app.storage.GetOrganizations()
+	if err != nil {
+		return nil, log.WrapActionError(log.ActionGet, model.TypeOrganization, nil, err)
+	}
+
+	return getOrganization, nil
+}
+
 func (app *application) admUpdateOrganization(ID string, name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) error {
 	err := app.storage.UpdateOrganization(ID, name, requestType, requiresOwnLogin, loginTypes, organizationDomains)
 	if err != nil {
@@ -215,4 +233,5 @@ func (app *application) admUpdateOrganization(ID string, name string, requestTyp
 	}
 
 	return err
+
 }