+++ driver/web/docs/gen/gen_types.go
@@ -126,7 +126,7 @@ type GlobalRole struct {
 type Organization struct {
 	Config           *OrganizationConfig `json:"config,omitempty"`
 	Id               string              `json:"id"`
-	LoginTypes       *[]string           `json:"login_types,omitempty"`
+	LoginTypes       *[]string           `json:"login_types"`
 	Name             string              `json:"name"`
 	RequiresOwnLogin *bool               `json:"requires_own_login,omitempty"`
 	Type             OrganizationType    `json:"type"`